feat: use FAB instead of old ugly button on transactions page
